--- src/pcss-civil-api/src/main/java/ca/bc/gov/open/pcss/pcsscivilapi/PcssCivilApiApplication.java
@@ -1,13 +1,0 @@
-package ca.bc.gov.open.pcss.pcsscivilapi;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PcssCivilApiApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(PcssCivilApiApplication.class, args);
-	}
-
-}
