--- src/pcss-civil-api/src/test/java/ca/bc/gov/open/pcss/pcsscivilapi/PcssCivilApiApplicationTests.java
@@ -1,13 +1,0 @@
-package ca.bc.gov.open.pcss.pcsscivilapi;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PcssCivilApiApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
