--- src/pcss-common-api/src/main/java/ca/bc/gov/open/pcss/pcsscommonapi/PcssCommonApiApplication.java
@@ -1,13 +1,0 @@
-package ca.bc.gov.open.pcss.pcsscommonapi;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PcssCommonApiApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(PcssCommonApiApplication.class, args);
-	}
-
-}
