--- src/pcss-common-api/src/test/java/ca/bc/gov/open/pcss/pcsscommonapi/PcssCommonApiApplicationTests.java
@@ -1,13 +1,0 @@
-package ca.bc.gov.open.pcss.pcsscomonapi;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PcssComonApiApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
