--- src/pcss-criminal-api/src/main/java/ca/bc/gov/open/pcss/pcsscriminalapi/PcssCriminalApiApplication.java
@@ -1,13 +1,0 @@
-package ca.bc.gov.open.pcss.pcsscriminalapi;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PcssCriminalApiApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(PcssCriminalApiApplication.class, args);
-	}
-
-}
