--- src/pcss-criminal-api/src/test/java/ca/bc/gov/open/pcss/pcsscriminalapi/PcssCriminalApiApplicationTests.java
@@ -1,13 +1,0 @@
-package ca.bc.gov.open.pcss.pcsscriminalapi;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PcssCriminalApiApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
