--- src/pcss-reports-api/src/main/java/ca/bc/gov/open/pcss/pcssreportsapi/PcssReportsApiApplication.java
@@ -1,13 +1,0 @@
-package ca.bc.gov.open.pcss.pcssreportsapi;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PcssReportsApiApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(PcssReportsApiApplication.class, args);
-	}
-
-}
