--- src/pcss-reports-api/src/test/java/ca/bc/gov/open/pcss/pcssreportsapi/PcssReportsApiApplicationTests.java
@@ -1,13 +1,0 @@
-package ca.bc.gov.open.pcss.pcssreportsapi;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PcssReportsApiApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
