+++ src/pcsswebservice/src/main/java/ca/bc/gov/open/pcss/pcsswebservice/PCSSWebServiceApplication.java
@@ -1,0 +1,13 @@
+package ca.bc.gov.open.pcss.pcsswebservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PCSSWebServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(PCSSWebServiceApplication.class, args);
+	}
+
+}
