+++ src/pcsswebservice/src/main/java/ca/bc/gov/open/pcss/pcsswebservice/WebServiceConfig.java
@@ -1,0 +1,84 @@
+package ca.bc.gov.open.pcss.pcsswebservice;
+
+import org.apache.cxf.Bus;
+import org.apache.cxf.bus.spring.SpringBus;
+import org.apache.cxf.jaxws.EndpointImpl;
+import org.apache.cxf.transport.servlet.CXFServlet;
+import org.springframework.boot.autoconfigure.web.servlet.DispatcherServletPath;
+import org.springframework.boot.web.servlet.ServletRegistrationBean;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Primary;
+
+import javax.xml.ws.Endpoint;
+
+
+@Configuration
+public class WebServiceConfig {
+
+    @Bean
+    public ServletRegistrationBean<CXFServlet> dispatcherServlet() {
+
+        return new ServletRegistrationBean(new CXFServlet(), "/ws/*");
+    }
+
+    @Bean
+    @Primary
+    public DispatcherServletPath dispatcherServletPathProvider() {
+
+        return () -> "";
+    }
+
+    @Bean(name=Bus.DEFAULT_BUS_ID)
+    public SpringBus springBus() {
+
+        SpringBus cxfBus = new SpringBus();
+        return cxfBus;
+    }
+
+    @Bean
+    public Endpoint CivilEndpoint(Bus bus, PcssCivilEndpoint pcssCivilEndpoint) {
+
+        EndpointImpl endpoint = new EndpointImpl(bus, pcssCivilEndpoint);
+        endpoint.publish("/pcssCivil");
+
+        return endpoint;
+    }
+
+    @Bean
+    public Endpoint CommonEndpoint(Bus bus, PcssCommonEndpoint pcssCommonEndpoint) {
+
+        EndpointImpl endpoint = new EndpointImpl(bus, pcssCommonEndpoint);
+        endpoint.publish("/pcssCommon");
+
+        return endpoint;
+    }
+
+    @Bean
+    public Endpoint CriminalEndpoint(Bus bus, PcssCriminalEndpoint pcssCriminalEndpoint) {
+
+        EndpointImpl endpoint = new EndpointImpl(bus, pcssCriminalEndpoint);
+        endpoint.publish("/pcssCriminal");
+
+        return endpoint;
+    }
+
+    @Bean
+    public Endpoint ReportEndpoint(Bus bus, PcssReportEndpoint pcssReportEndpoint) {
+
+        EndpointImpl endpoint = new EndpointImpl(bus, pcssReportEndpoint);
+        endpoint.publish("/pcssReport");
+
+        return endpoint;
+    }
+
+    @Bean
+    public Endpoint JusticeCddsEndpoint(Bus bus, JusticeCddsEndpoint justiceCddsEndpoint) {
+
+        EndpointImpl endpoint = new EndpointImpl(bus, justiceCddsEndpoint);
+        endpoint.publish("/justiceCDDS");
+
+        return endpoint;
+    }
+
+}
