+++ src/pcsswebservice/src/test/java/ca/bc/gov/open/pcss/pcsswebservice/PCSSWebServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package ca.bc.gov.open.pcss.pcsswebservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class PCSSWebServiceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
